Move token.js to lib directory
