+++ src/lib/token.js
